requirements.txt
dill
gdown
transformers
flask
torch
joblib
Pillow
